requirements.txt
# 基础依赖
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.3.0
PyYAML>=5.4.0
tqdm>=4.60.0
requests>=2.25.0
scipy>=1.7.0
pandas>=1.3.0

# YOLO框架
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0

# 标注工具
labelImg>=1.8.0
roboflow>=1.0.0

# 可视化工具
seaborn>=0.11.0
plotly>=5.0.0
tensorboard>=2.7.0
wandb>=0.12.0
